venv ignored. ticker normalization. continuous polling
venv should not be in the git. added to git ignore to prevent
reoccurences.

normalized ticker in watchlist to reduce errors.

continuous polling allows us to add to watchlist without reload. We also
replace outterhtmls with innerhtmls because no element is being removed,
but rather an element is being updated (it is getting created withing
the card, but this innerHTML of the card takes priority).